activation of acceptance and integration tests

+++ test/acceptance/plugin.acceptance.test.ts
@@ -1,26 +1,74 @@
 import execa from 'execa'
-import * as fs from 'fs-extra'
+import fs from 'fs-extra'
+import {fileURLToPath} from 'node:url'
 import * as path from 'path'
 
-const plugins = ['heroku-ps-exec']
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+const plugins = ['@heroku-cli/plugin-applink']
 
 const skipOnWindows = process.platform === 'win32' ? it.skip : it
 
-describe.skip('plugins', function () {
+function resolvePluginPath(plugin: string): null | string {
+  const candidates = [
+    path.resolve(__dirname, '../../node_modules', plugin, 'package.json'),
+    path.resolve(__dirname, '../../../node_modules', plugin, 'package.json'),
+  ]
+
+  for (const p of candidates) {
+    if (fs.existsSync(p)) return path.dirname(p)
+  }
+
+  return null
+}
+
+function safeCloneDirName(plugin: string): string {
+  return plugin.replace('@heroku-cli/', '')
+}
+
+describe('plugins', function () {
   plugins.forEach(plugin => {
+    const pluginRoot = resolvePluginPath(plugin)
+    if (!pluginRoot) {
+      it.skip(plugin, async function () {})
+      return
+    }
+
     skipOnWindows(plugin, async () => {
-      const cwd = path.join(__dirname, '../../tmp/plugin', plugin)
+      const cwd = path.resolve(__dirname, '../../tmp/plugin', safeCloneDirName(plugin))
       await fs.remove(cwd)
-      const pkg = await fs.readJSON(path.join(__dirname, '../../node_modules', plugin, 'package.json'))
-      if (!pkg.repository) {
+      const pkg = await fs.readJSON(path.join(pluginRoot, 'package.json'))
+      const repo = pkg.repository
+
+      if (!repo) {
         throw new Error('No repository found')
       }
 
-      await execa('git', ['clone', pkg.repository.url.split('+')[1], cwd])
+      const repoUrl = typeof repo === 'string' ? repo : repo.url
+
+      if (!repoUrl) {
+        throw new Error('No repository URL found in package.json')
+      }
+
+      let cloneUrl: string
+
+      if (repoUrl.includes('+')) {
+        cloneUrl = repoUrl.split('+')[1]
+      } else if (repoUrl.startsWith('github:')) {
+        cloneUrl = `https://github.com/${repoUrl.slice(7)}.git`
+      } else if (repoUrl.startsWith('http://') || repoUrl.startsWith('https://')) {
+        cloneUrl = repoUrl
+      } else if (/^[^/]+\/[^/]+$/.test(repoUrl)) {
+        cloneUrl = `https://github.com/${repoUrl}.git`
+      } else {
+        cloneUrl = repoUrl
+      }
+
+      await execa('git', ['clone', cloneUrl, cwd])
       const opts = {cwd, stdio: [0, 1, 2]}
       await execa('git', ['checkout', `v${pkg.version}`], opts)
-      await execa('yarn', [], opts)
-      await execa('yarn', ['test'], opts)
+      await execa('npm', [], opts)
+      await execa('npm', ['test'], opts)
     })
   })
 })
